Remove OLS operator code and RBAC rules
Remove controller logic (funcs.go, ols_install.go,
openstacklightspeed_controller.go) and RBAC permissions (OLSConfig
access, OLM operator resources) that are no longer needed.

This prepares the operator for migration to lightspeed-core by
stripping out all OLS-specific functionality.
