+++ config/rbac/role.yaml
@@ -38,74 +38,3 @@ rules:
   - get
   - patch
   - update
-- apiGroups:
-  - ols.openshift.io
-  resources:
-  - olsconfigs
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
-- apiGroups:
-  - ols.openshift.io
-  resources:
-  - olsconfigs/finalizers
-  verbs:
-  - update
-- apiGroups:
-  - ols.openshift.io
-  resources:
-  - olsconfigs/status
-  verbs:
-  - get
-  - patch
-  - update
-- apiGroups:
-  - operators.coreos.com
-  resources:
-  - clusterserviceversions
-  verbs:
-  - get
-  - list
-  - watch
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: Role
-metadata:
-  name: manager-role
-  namespace: openshift-lightspeed
-rules:
-- apiGroups:
-  - operators.coreos.com
-  resources:
-  - clusterserviceversions
-  verbs:
-  - delete
-  - patch
-  - update
-- apiGroups:
-  - operators.coreos.com
-  resources:
-  - installplans
-  verbs:
-  - delete
-  - get
-  - list
-  - update
-  - watch
-- apiGroups:
-  - operators.coreos.com
-  resources:
-  - subscriptions
-  verbs:
-  - create
-  - delete
-  - get
-  - list
-  - patch
-  - update
-  - watch
