+++ internal/controller/funcs.go
@@ -17,23 +17,13 @@ limitations under the License.
 package controller
 
 import (
-	"context"
-	"fmt"
-	"math/rand"
-	"strconv"
-
-	apiv1beta1 "github.com/openstack-lightspeed/operator/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
 
 	_ "embed"
 
 	common_helper "github.com/openstack-k8s-operators/lib-common/modules/common/helper"
-	k8s_errors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	uns "k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
 const (
@@ -56,258 +46,6 @@ const (
 	// OLSConfigName - OLS forbids other name for OLSConfig instance than OLSConfigName
 	OLSConfigName = "cluster"
 )
-
-// systemPrompt - system prompt tailored to the needs of OpenStack Lightspeed. It overwrites the default OLS prompt.
-//
-//go:embed system_prompt.txt
-var systemPrompt string
-
-// GetSystemPrompt returns the OpenStackLightspeed system prompt
-func GetSystemPrompt() string {
-	return systemPrompt
-}
-
-// RemoveOLSConfig attempts to remove the OLSConfig custom resource if it exists
-// and is managed by the given OpenStackLightspeed instance. It first fetches the OLSConfig,
-// checks whether the current OpenStackLightspeed instance is the owner (via label check),
-// and if so, removes the finalizer and deletes the OLSConfig resource.
-// Returns (true, nil) if the OLSConfig is not found (indicating it has already been deleted).
-// Returns (true, nil) if the resource was deleted successfully, or (false, error) if any error occurs.
-func RemoveOLSConfig(
-	ctx context.Context,
-	helper *common_helper.Helper,
-	instance *apiv1beta1.OpenStackLightspeed,
-) (bool, error) {
-	olsConfig, err := GetOLSConfig(ctx, helper)
-	if err != nil && !k8s_errors.IsNotFound(err) {
-		return false, err
-	} else if err != nil && k8s_errors.IsNotFound(err) {
-		return true, nil
-	}
-
-	_, err = controllerutil.CreateOrPatch(ctx, helper.GetClient(), &olsConfig, func() error {
-		ownerLabel := olsConfig.GetLabels()[OpenStackLightspeedOwnerIDLabel]
-		isInstanceOwnedOLSConfig := ownerLabel == string(instance.GetObjectMeta().GetUID())
-
-		if ownerLabel == "" || !isInstanceOwnedOLSConfig {
-			helper.GetLogger().Info("Skipping OLSConfig deletion as it is not managed by the OpenStackLightspeed instance")
-			return nil
-		}
-
-		if ok := controllerutil.RemoveFinalizer(&olsConfig, helper.GetFinalizer()); !ok {
-			return fmt.Errorf("remove finalizer failed")
-		}
-
-		return nil
-	})
-	if err != nil {
-		return false, err
-	}
-
-	err = helper.GetClient().Delete(ctx, &olsConfig)
-	if err != nil {
-		return false, err
-	}
-
-	_, err = GetOLSConfig(ctx, helper)
-	if err != nil && k8s_errors.IsNotFound(err) {
-		return true, nil
-	} else if err != nil {
-		return false, err
-	}
-
-	return false, nil
-}
-
-// GetOLSConfig returns OLSConfig if there is one present in the cluster.
-func GetOLSConfig(ctx context.Context, helper *common_helper.Helper) (uns.Unstructured, error) {
-	OLSConfigGVR := schema.GroupVersionResource{
-		Group:    "ols.openshift.io",
-		Version:  "v1alpha1",
-		Resource: "olsconfigs",
-	}
-
-	OLSConfigList := &uns.UnstructuredList{}
-	OLSConfigList.SetGroupVersionKind(OLSConfigGVR.GroupVersion().WithKind("OLSConfig"))
-	err := helper.GetClient().List(ctx, OLSConfigList)
-	if err != nil {
-		return uns.Unstructured{}, err
-	}
-
-	if len(OLSConfigList.Items) > 0 {
-		return OLSConfigList.Items[0], nil
-	}
-
-	return uns.Unstructured{}, k8s_errors.NewNotFound(
-		schema.GroupResource{Group: "ols.openshifg.io", Resource: "olsconfigs"},
-		"OLSConfig")
-}
-
-// BuildRAGConfigs builds the RAG configuration array.
-// OpenStack RAG is always included first.
-// OCP RAG is added if ocpVersion is provided.
-func BuildRAGConfigs(instance *apiv1beta1.OpenStackLightspeed, ocpVersion string) []interface{} {
-	rags := []interface{}{
-		// OpenStack RAG
-		map[string]interface{}{
-			"image":     instance.Spec.RAGImage,
-			"indexPath": OpenStackLightspeedVectorDBPath,
-		},
-	}
-
-	// Add OCP RAG if enabled
-	if ocpVersion != "" {
-		rags = append(rags, map[string]interface{}{
-			"image":     instance.Spec.RAGImage,
-			"indexPath": GetOCPVectorDBPath(ocpVersion),
-			"indexID":   GetOCPIndexName(ocpVersion),
-		})
-	}
-
-	return rags
-}
-
-// PatchOLSConfig patches OLSConfig with information from OpenStackLightspeed instance.
-func PatchOLSConfig(
-	helper *common_helper.Helper,
-	instance *apiv1beta1.OpenStackLightspeed,
-	olsConfig *uns.Unstructured,
-) error {
-	// Patch the Providers section
-	providersPatch := []interface{}{
-		map[string]interface{}{
-			"credentialsSecretRef": map[string]interface{}{
-				"name": instance.Spec.LLMCredentials,
-			},
-			"models": []interface{}{
-				map[string]interface{}{
-					"name": instance.Spec.ModelName,
-					"parameters": map[string]interface{}{
-						"maxTokensForResponse": float64(instance.Spec.MaxTokensForResponse), // unstructured JSON numbers default to float64
-					},
-				},
-			},
-			"name": OpenStackLightspeedDefaultProvider,
-			"type": instance.Spec.LLMEndpointType,
-			"url":  instance.Spec.LLMEndpoint,
-		},
-	}
-
-	provider := providersPatch[0].(map[string]interface{})
-	if instance.Spec.LLMProjectID != "" {
-		if err := uns.SetNestedField(provider, instance.Spec.LLMProjectID, "projectID"); err != nil {
-			return err
-		}
-	}
-
-	if instance.Spec.LLMDeploymentName != "" {
-		if err := uns.SetNestedField(provider, instance.Spec.LLMDeploymentName, "deploymentName"); err != nil {
-			return err
-		}
-	}
-
-	if instance.Spec.LLMAPIVersion != "" {
-		if err := uns.SetNestedField(provider, instance.Spec.LLMAPIVersion, "apiVersion"); err != nil {
-			return err
-		}
-	}
-
-	if err := uns.SetNestedSlice(olsConfig.Object, providersPatch, "spec", "llm", "providers"); err != nil {
-		return err
-	}
-
-	// Patch the RAG section
-	// Build RAG array with priorities using BuildRAGConfigs
-	ragConfigs := BuildRAGConfigs(instance, instance.Status.ActiveOCPRAGVersion)
-
-	if err := uns.SetNestedSlice(olsConfig.Object, ragConfigs, "spec", "ols", "rag"); err != nil {
-		return err
-	}
-
-	if instance.Spec.TLSCACertBundle != "" {
-		tlsCaCertBundle := instance.Spec.TLSCACertBundle
-		err := uns.SetNestedField(olsConfig.Object, tlsCaCertBundle, "spec", "ols", "additionalCAConfigMapRef", "name")
-		if err != nil {
-			return err
-		}
-	}
-
-	modelName := instance.Spec.ModelName
-	err := uns.SetNestedField(olsConfig.Object, modelName, "spec", "ols", "defaultModel")
-	if err != nil {
-		return err
-	}
-
-	err = uns.SetNestedField(olsConfig.Object, OpenStackLightspeedDefaultProvider, "spec", "ols", "defaultProvider")
-	if err != nil {
-		return err
-	}
-
-	// Disable the OCP RAG
-	// TODO(lucasagomes): Remove this once we have a "query router" that can
-	// handle multiple RAGs nicely
-	err = uns.SetNestedField(olsConfig.Object, true, "spec", "ols", "byokRAGOnly")
-	if err != nil {
-		return err
-	}
-
-	// Disable or enable feedback collection
-	err = uns.SetNestedField(olsConfig.Object, instance.Spec.FeedbackDisabled, "spec", "ols", "userDataCollection", "feedbackDisabled")
-	if err != nil {
-		return err
-	}
-
-	// Disable or enable transcripts collection
-	err = uns.SetNestedField(olsConfig.Object, instance.Spec.TranscriptsDisabled, "spec", "ols", "userDataCollection", "transcriptsDisabled")
-	if err != nil {
-		return err
-	}
-
-	err = uns.SetNestedField(olsConfig.Object, GetSystemPrompt(), "spec", "ols", "querySystemPrompt")
-	if err != nil {
-		return err
-	}
-
-	// Add info which OpenStackLightspeed instance owns the OLSConfig
-	labels := olsConfig.GetLabels()
-	updatedLabels := map[string]interface{}{
-		OpenStackLightspeedOwnerIDLabel: string(instance.GetUID()),
-	}
-	for k, v := range labels {
-		updatedLabels[k] = v
-	}
-
-	err = uns.SetNestedField(olsConfig.Object, updatedLabels, "metadata", "labels")
-	if err != nil {
-		return err
-	}
-
-	// Add OpenStack finalizers
-	if !controllerutil.AddFinalizer(olsConfig, helper.GetFinalizer()) && instance.Status.Conditions == nil {
-		return fmt.Errorf("cannot add finalizer")
-	}
-
-	return nil
-}
-
-// IsOLSConfigReady returns true if OLSConfig's overallStatus is Ready
-func IsOLSConfigReady(ctx context.Context, helper *common_helper.Helper) (bool, error) {
-	olsConfig, err := GetOLSConfig(ctx, helper)
-	if err != nil {
-		return false, err
-	}
-
-	overallStatus, found, err := uns.NestedString(olsConfig.Object, "status", "overallStatus")
-	if err != nil {
-		return false, err
-	}
-
-	if !found || overallStatus != "Ready" {
-		return false, OLSConfigPing(ctx, helper)
-	}
-
-	return true, nil
-}
 
 // IsOwnedBy returns true if 'object' is owned by 'owner' based on OwnerReference UID.
 func IsOwnedBy(object metav1.Object, owner metav1.Object) bool {
@@ -334,31 +72,4 @@ func GetRawClient(helper *common_helper.Helper) (client.Client, error) {
 	}
 
 	return rawClient, nil
-}
-
-// OLSConfigPing adds a random label to the OLSConfig to trigger a reconciliation
-// by the OpenShift Lightspeed operator. This causes the operator to update the Status field.
-// Note: This is a workaround for a current limitation—when the OLS operator is installed
-// in the openstack-lightspeed namespace, it does not automatically update the OLSConfig
-// status as expected.
-func OLSConfigPing(ctx context.Context, helper *common_helper.Helper) error {
-	const randomLabelKey = "openstack-lightspeed/ping"
-
-	olsConfig, err := GetOLSConfig(ctx, helper)
-	if err != nil {
-		return err
-	}
-
-	labels := olsConfig.GetLabels()
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-
-	labels[randomLabelKey] = strconv.Itoa(rand.Int())
-	olsConfig.SetLabels(labels)
-
-	if err := helper.GetClient().Update(ctx, &olsConfig); err != nil {
-		return err
-	}
-	return nil
 }
