+++ internal/controller/openstacklightspeed_controller.go
@@ -19,7 +19,6 @@ package controller
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/go-logr/logr"
 	"github.com/openstack-k8s-operators/lib-common/modules/common/condition"
@@ -56,14 +55,6 @@ func (r *OpenStackLightspeedReconciler) GetLogger(ctx context.Context) logr.Logg
 // +kubebuilder:rbac:groups=lightspeed.openstack.org,resources=openstacklightspeeds,verbs=get;list;watch;create;update;patch;delete
 // +kubebuilder:rbac:groups=lightspeed.openstack.org,resources=openstacklightspeeds/status,verbs=get;update;patch
 // +kubebuilder:rbac:groups=lightspeed.openstack.org,resources=openstacklightspeeds/finalizers,verbs=update
-// +kubebuilder:rbac:groups=ols.openshift.io,resources=olsconfigs,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=ols.openshift.io,resources=olsconfigs/status,verbs=get;update;patch
-// +kubebuilder:rbac:groups=ols.openshift.io,resources=olsconfigs/finalizers,verbs=update
-// +kubebuilder:rbac:groups=operators.coreos.com,resources=clusterserviceversions,verbs=get;list;watch
-// +kubebuilder:rbac:groups=operators.coreos.com,resources=clusterserviceversions,namespace=openshift-lightspeed,verbs=update;patch;delete
-// +kubebuilder:rbac:groups=operators.coreos.com,resources=subscriptions,namespace=openshift-lightspeed,verbs=get;list;watch;create;update;patch;delete
-// +kubebuilder:rbac:groups=operators.coreos.com,resources=installplans,namespace=openshift-lightspeed,verbs=get;list;watch;update;delete
-// +kubebuilder:rbac:groups=config.openshift.io,resources=clusterversions,verbs=get;list;watch
 
 // Reconcile is part of the main kubernetes reconciliation loop which aims to
 // move the current state of the cluster closer to the desired state.
@@ -147,7 +138,8 @@ func (r *OpenStackLightspeedReconciler) Reconcile(ctx context.Context, req ctrl.
 	r.resolveOCPVersion(ctx, helper, instance)
 
 	if !instance.DeletionTimestamp.IsZero() {
-		return r.reconcileDelete(ctx, helper, instance)
+		r.reconcileDelete(ctx, helper, instance)
+		return ctrl.Result{}, nil
 	}
 
 	if instance.DeletionTimestamp.IsZero() && controllerutil.AddFinalizer(instance, helper.GetFinalizer()) {
@@ -160,98 +152,6 @@ func (r *OpenStackLightspeedReconciler) Reconcile(ctx context.Context, req ctrl.
 
 	if instance.Spec.MaxTokensForResponse == 0 {
 		instance.Spec.MaxTokensForResponse = apiv1beta1.OpenStackLightspeedDefaultValues.MaxTokensForResponse
-	}
-
-	// Ensure a compatible version of the OpenShift Lightspeed Operator is running in the cluster.
-	// This checks if the correct OLS Operator version is present and installs it if necessary.
-	isOLSOperatorInstalled, err := EnsureOLSOperatorInstalled(ctx, helper, instance)
-	if err != nil {
-		instance.Status.Conditions.Set(condition.FalseCondition(
-			apiv1beta1.OpenShiftLightspeedOperatorReadyCondition,
-			condition.ErrorReason,
-			condition.SeverityWarning,
-			condition.DeploymentReadyErrorMessage,
-			err.Error(),
-		))
-
-		return ctrl.Result{}, nil
-	} else if !isOLSOperatorInstalled {
-		instance.Status.Conditions.Set(condition.FalseCondition(
-			apiv1beta1.OpenShiftLightspeedOperatorReadyCondition,
-			condition.RequestedReason,
-			condition.SeverityInfo,
-			apiv1beta1.OpenShiftLightspeedOperatorWaiting,
-		))
-
-		// In this branch we know that the
-		return ctrl.Result{Requeue: true, RequeueAfter: 10 * time.Second}, nil
-	}
-
-	// Mark the OpenShift Lightspeed Operator as ready in the status conditions.
-	instance.Status.Conditions.MarkTrue(
-		apiv1beta1.OpenShiftLightspeedOperatorReadyCondition,
-		apiv1beta1.OpenShiftLightspeedOperatorReady,
-	)
-
-	// NOTE: We cannot consume the OLSConfig definition directly from the OLS operator's code due to
-	// a conflict in Go versions. When this comment was written, the min. required Go version for
-	// openstack-operator was 1.21 whereas OLS operator required at least Go version 1.23. Once the
-	// Go versions catch up with each other we should consider consuming OLSConfig directly from OLS
-	// operator and updating this code and any subsequent code that consumes this structure.
-	olsConfig := uns.Unstructured{}
-	olsConfigGVK := schema.GroupVersionKind{
-		Group:   "ols.openshift.io",
-		Version: "v1alpha1",
-		Kind:    "OLSConfig",
-	}
-
-	olsConfig.SetGroupVersionKind(olsConfigGVK)
-	olsConfig.SetName(OLSConfigName)
-
-	_, err = controllerutil.CreateOrPatch(ctx, r.Client, &olsConfig, func() error {
-		// Check if the OpenStackLightspeed instance that is being processed owns the OLSConfig. If
-		// it is owned by other OpenStackLightspeed instance stop the reconciliation.
-		olsConfigLabels := olsConfig.GetLabels()
-		ownerLabel := ""
-		if val, ok := olsConfigLabels[OpenStackLightspeedOwnerIDLabel]; ok {
-			ownerLabel = val
-		}
-
-		if ownerLabel != "" && ownerLabel != string(instance.GetObjectMeta().GetUID()) {
-			return fmt.Errorf("OLSConfig is managed by different OpenStackLightspeed instance")
-		}
-
-		err = PatchOLSConfig(helper, instance, &olsConfig)
-		if err != nil {
-			return err
-		}
-
-		return nil
-	})
-	if err != nil {
-		instance.Status.Conditions.Set(condition.FalseCondition(
-			apiv1beta1.OpenStackLightspeedReadyCondition,
-			condition.ErrorReason,
-			condition.SeverityWarning,
-			condition.DeploymentReadyErrorMessage,
-			err.Error()))
-		return ctrl.Result{}, err
-	}
-
-	OLSConfigReady, err := IsOLSConfigReady(ctx, helper)
-	if err != nil {
-		return ctrl.Result{}, err
-	}
-
-	if OLSConfigReady {
-		instance.Status.Conditions.MarkTrue(
-			apiv1beta1.OpenStackLightspeedReadyCondition,
-			apiv1beta1.OpenStackLightspeedReadyMessage,
-		)
-		Log.Info("OLSConfig is ready!")
-	} else {
-		Log.Info("OLSConfig is not ready yet. Waiting...")
-		return ctrl.Result{RequeueAfter: time.Second * time.Duration(5)}, nil
 	}
 
 	Log.Info("OpenStackLightspeed Reconciled successfully")
@@ -351,30 +251,13 @@ func (r *OpenStackLightspeedReconciler) reconcileDelete(
 	ctx context.Context,
 	helper *common_helper.Helper,
 	instance *apiv1beta1.OpenStackLightspeed,
-) (ctrl.Result, error) {
+) {
 	Log := r.GetLogger(ctx)
 	Log.Info("OpenStackLightspeed Reconciling Delete")
-
-	isRemoved, err := RemoveOLSConfig(ctx, helper, instance)
-	if err != nil {
-		return ctrl.Result{}, err
-	} else if !isRemoved {
-		Log.Info("OLSConfig removal in progress ...")
-		return ctrl.Result{RequeueAfter: time.Second * 10}, nil
-	}
-
-	isUninstalled, err := UninstallInstanceOwnedOLSOperator(ctx, helper, instance)
-	if err != nil {
-		return ctrl.Result{}, err
-	} else if !isUninstalled {
-		Log.Info("OLS Operator uninstallation in progress ...")
-		return ctrl.Result{RequeueAfter: time.Second * 10}, nil
-	}
 
 	controllerutil.RemoveFinalizer(instance, helper.GetFinalizer())
 
 	Log.Info("OpenStackLightspeed Reconciling Delete completed")
-	return ctrl.Result{}, nil
 }
 
 // SetupWithManager sets up the controller with the Manager.
