+++ scripts/env.sh
@@ -1,3 +1,2 @@
 #!/bin/bash
-export OPENSHIFT_LIGHTSPEED_OPERATOR_VERSION="latest"
 export WATCH_NAMESPACE="openshift-lightspeed"
